Simple container to sync start time between multiple containers
While testing network throughput with multiple containers it was necessary to make
sure tests started together.

This is a simple approach with a python server that receives a post  and blocks the
caller until a certain number of containers do the same, then unblocks the all the
callers at once and tests can start almost simultaneously.

Signed-off-by: Mauro S. M. Rodrigues <maurosr@linux.vnet.ibm.com>